Safari-609.1.11.1/LayoutTests/crypto/workers/subtle/pbkdf2-import-key-expected.txt

[Worker] Test importing a PBKDF2 raw key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/pbkdf2-import-key.js
PASS [Worker] publicKey.toString() is '[object CryptoKey]'
PASS [Worker] publicKey.type is 'secret'
PASS [Worker] publicKey.extractable is false
PASS [Worker] publicKey.algorithm.name is 'PBKDF2'
PASS [Worker] publicKey.usages is ['deriveBits', 'deriveKey']
PASS successfullyParsed is true

TEST COMPLETE
